Split dns_server into a module per class
