--- sphinx-doc/infrastructure/dns_server.rst
@@ -1,5 +1,0 @@
-
-.. automodule:: infrastructure.dns_server
-   :special-members: __init__
-   :members:
-   :member-order: bysource
