+++ sphinx-doc/infrastructure/dns_server/index.rst
@@ -1,0 +1,10 @@
+DNS Server
+==========
+
+Contents:
+
+.. toctree::
+   :maxdepth: 2
+   :glob:
+
+   *
