+++ sphinx-doc/infrastructure/dns_server/logger.rst
@@ -1,0 +1,5 @@
+
+.. automodule:: infrastructure.dns_server.logger
+   :special-members: __init__
+   :members:
+
