+++ sphinx-doc/infrastructure/dns_server/main.rst
@@ -1,0 +1,5 @@
+
+.. automodule:: infrastructure.dns_server.main
+   :special-members: __init__
+   :members:
+
