+++ sphinx-doc/infrastructure/dns_server/resolver.rst
@@ -1,0 +1,5 @@
+
+.. automodule:: infrastructure.dns_server.resolver
+   :special-members: __init__
+   :members:
+
