+++ sphinx-doc/infrastructure/dns_server/servers.rst
@@ -1,0 +1,5 @@
+
+.. automodule:: infrastructure.dns_server.servers
+   :special-members: __init__
+   :members:
+
