+++ sphinx-doc/infrastructure/index.rst
@@ -9,3 +9,4 @@ Contents:
 
    *
    beacon_server/index
+   dns_server/index
